+++ src/reducers/isLoading.js
@@ -1,0 +1,5 @@
+export const loading = (state = false, action) => {
+  switch(action.type) {
+    case ''
+  }
+}
